+++ kotlinpoet-elementhandler-elements/src/main/kotlin/com/squareup/kotlinpoet/elementhandler/elements/ElementsElementHandler.kt
@@ -7,16 +7,16 @@ import com.google.common.collect.LinkedHashMultimap
 import com.google.common.collect.SetMultimap
 import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.CodeBlock
-import com.squareup.kotlinpoet.km.ImmutableKmClass
-import com.squareup.kotlinpoet.km.KotlinPoetKm
-import com.squareup.kotlinpoet.km.specs.ElementHandler
-import com.squareup.kotlinpoet.km.specs.ElementHandler.JvmFieldModifier
-import com.squareup.kotlinpoet.km.specs.ElementHandler.JvmFieldModifier.TRANSIENT
-import com.squareup.kotlinpoet.km.specs.ElementHandler.JvmFieldModifier.VOLATILE
-import com.squareup.kotlinpoet.km.specs.ElementHandler.JvmMethodModifier
-import com.squareup.kotlinpoet.km.specs.ElementHandler.JvmMethodModifier.STATIC
-import com.squareup.kotlinpoet.km.specs.ElementHandler.JvmMethodModifier.SYNCHRONIZED
-import com.squareup.kotlinpoet.km.toImmutableKmClass
+import com.squareup.kotlinpoet.metadata.ImmutableKmClass
+import com.squareup.kotlinpoet.metadata.KotlinPoetKm
+import com.squareup.kotlinpoet.metadata.specs.ElementHandler
+import com.squareup.kotlinpoet.metadata.specs.ElementHandler.JvmFieldModifier
+import com.squareup.kotlinpoet.metadata.specs.ElementHandler.JvmFieldModifier.TRANSIENT
+import com.squareup.kotlinpoet.metadata.specs.ElementHandler.JvmFieldModifier.VOLATILE
+import com.squareup.kotlinpoet.metadata.specs.ElementHandler.JvmMethodModifier
+import com.squareup.kotlinpoet.metadata.specs.ElementHandler.JvmMethodModifier.STATIC
+import com.squareup.kotlinpoet.metadata.specs.ElementHandler.JvmMethodModifier.SYNCHRONIZED
+import com.squareup.kotlinpoet.metadata.toImmutableKmClass
 import kotlinx.metadata.jvm.JvmFieldSignature
 import kotlinx.metadata.jvm.JvmMethodSignature
 import java.util.concurrent.ConcurrentHashMap
