--- kotlinpoet-km-specs/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=kotlinpoet-km-specs
-POM_NAME=KotlinPoet (KotlinPoet-km specs interop)
-POM_DESCRIPTION=Extensions for generating KotlinPoet specs from KotlinPoet-km types.
-POM_PACKAGING=jar
