--- kotlinpoet-km-specs/src/main/kotlin/com/squareup/kotlinpoet/km/specs/ElementHandler.kt
@@ -1,162 +1,0 @@
-/*
- * Copyright (C) 2019 Square, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.squareup.kotlinpoet.km.specs
-
-import com.squareup.kotlinpoet.AnnotationSpec
-import com.squareup.kotlinpoet.CodeBlock
-import com.squareup.kotlinpoet.asClassName
-import com.squareup.kotlinpoet.km.ImmutableKmClass
-import com.squareup.kotlinpoet.km.KotlinPoetKm
-import kotlinx.metadata.jvm.JvmFieldSignature
-import kotlinx.metadata.jvm.JvmMethodSignature
-
-/**
- * A basic interface for looking up information about JVM elements.
- */
-@KotlinPoetKm
-interface ElementHandler {
-
-  interface JvmModifier {
-    fun annotationSpec(): AnnotationSpec
-  }
-
-  /** Modifiers that are annotations in Kotlin but modifier keywords in bytecode. */
-  enum class JvmFieldModifier : JvmModifier {
-    STATIC {
-      override fun annotationSpec(): AnnotationSpec = AnnotationSpec.builder(JvmStatic::class.asClassName()).build()
-    },
-    TRANSIENT {
-      override fun annotationSpec(): AnnotationSpec = AnnotationSpec.builder(Transient::class.asClassName()).build()
-    },
-    VOLATILE {
-      override fun annotationSpec(): AnnotationSpec = AnnotationSpec.builder(Volatile::class.asClassName()).build()
-    };
-  }
-
-  /** Modifiers that are annotations in Kotlin but modifier keywords in bytecode. */
-  enum class JvmMethodModifier : JvmModifier {
-    STATIC {
-      override fun annotationSpec(): AnnotationSpec = AnnotationSpec.builder(JvmStatic::class.asClassName()).build()
-    },
-    SYNCHRONIZED {
-      override fun annotationSpec(): AnnotationSpec = AnnotationSpec.builder(Synchronized::class.asClassName()).build()
-    }
-  }
-
-  /**
-   * Indicates if this element handler supports [AnnotationRetention.RUNTIME]-retained annotations.
-   * This is used to indicate if manual inference of certain non-RUNTIME-retained annotations should
-   * be done, such as [JvmName].
-   */
-  val supportsNonRuntimeRetainedAnnotations: Boolean
-
-  /**
-   * Looks up other classes, such as for nested members. Note that this class would always be
-   * Kotlin, so Metadata can be relied on for this.
-   *
-   * @param jvmName The JVM name of the class (example: `"org/foo/bar/Baz$Nested"`).
-   * @return the read [ImmutableKmClass] from its metadata. If no class was found, this should throw
-   *         an exception.
-   */
-  fun classFor(jvmName: String): ImmutableKmClass
-
-  /**
-   * Looks up a class and returns whether or not it is an interface. Note that this class can be
-   * Java or Kotlin, so Metadata should not be relied on for this.
-   *
-   * @param jvmName The JVM name of the class (example: `"org/foo/bar/Baz$Nested"`).
-   * @return whether or not it is an interface.
-   */
-  fun isInterface(jvmName: String): Boolean
-
-  /**
-   * Looks up a given class field given its [JvmFieldSignature] and returns any [JvmModifier]s
-   * found on it.
-   *
-   * @param classJvmName The JVM name of the class (example: `"org/foo/bar/Baz$Nested"`).
-   * @param fieldSignature The field to look up.
-   * @return the set of found modifiers.
-   */
-  fun fieldJvmModifiers(classJvmName: String, fieldSignature: JvmFieldSignature): Set<JvmFieldModifier>
-
-  /**
-   * Looks up the annotations on a given class field given its [JvmFieldSignature].
-   *
-   * @param classJvmName The JVM name of the class (example: `"org/foo/bar/Baz$Nested"`).
-   * @param fieldSignature The field with annotations to look up.
-   * @return the [AnnotationSpec] representations of the annotations on the target field.
-   */
-  fun fieldAnnotations(classJvmName: String, fieldSignature: JvmFieldSignature): List<AnnotationSpec>
-
-  /**
-   * Looks up a given class method given its [JvmMethodSignature] and returns any [JvmMethodModifier]s
-   * found on it.
-   *
-   * @param classJvmName The JVM name of the class (example: `"org/foo/bar/Baz$Nested"`).
-   * @param methodSignature The method with annotations to look up.
-   * @return the set of found modifiers.
-   */
-  fun methodJvmModifiers(classJvmName: String, methodSignature: JvmMethodSignature): Set<JvmMethodModifier>
-
-  /**
-   * Looks up the annotations on a given class constructor given its [JvmMethodSignature].
-   *
-   * @param classJvmName The JVM name of the class (example: `"org/foo/bar/Baz$Nested"`).
-   * @param constructorSignature The constructor with annotations to look up.
-   * @return the [AnnotationSpec] representations of the annotations on the target method.
-   */
-  fun constructorAnnotations(classJvmName: String, constructorSignature: JvmMethodSignature): List<AnnotationSpec>
-
-  /**
-   * Looks up the annotations on a given class method given its [JvmMethodSignature].
-   *
-   * @param classJvmName The JVM name of the class (example: `"org/foo/bar/Baz$Nested"`).
-   * @param methodSignature The method with annotations to look up.
-   * @return the [AnnotationSpec] representations of the annotations on the target method.
-   */
-  fun methodAnnotations(classJvmName: String, methodSignature: JvmMethodSignature): List<AnnotationSpec>
-
-  /**
-   * Looks up the enum entry on a given enum given its member name.
-   *
-   * @param enumClassJvmName The JVM name of the enum class (example: `"org/foo/bar/Baz$Nested"`).
-   * @param memberName The simple member name.
-   * @return the read [ImmutableKmClass] from its metadata, if any. For simple enum members with no
-   *         class bodies, this should always be null.
-   */
-  fun enumEntry(enumClassJvmName: String, memberName: String): ImmutableKmClass?
-
-  /**
-   * Looks up the constant value on a given class field given its [JvmFieldSignature].
-   *
-   * @param classJvmName The JVM name of the class (example: `"org/foo/bar/Baz$Nested"`).
-   * @param fieldSignature The field with annotations to look up.
-   * @return the [CodeBlock] representation of the constant on the target field. Null if the value
-   *         cannot be resolved.
-   */
-  fun fieldConstant(classJvmName: String, fieldSignature: JvmFieldSignature): CodeBlock?
-
-  /**
-   * Looks up if a given [methodSignature] within [classJvmName] is an override of another method.
-   * Implementers should search for a matching method signature in the supertypes/classes of
-   * [classJvmName] to see if there are matches.
-   *
-   * @param classJvmName The JVM name of the class (example: `"org/foo/bar/Baz$Nested"`).
-   * @param methodSignature The method signature to check.
-   * @return whether or not the method is an override.
-   */
-  fun isMethodOverride(classJvmName: String, methodSignature: JvmMethodSignature): Boolean
-}
