--- kotlinpoet-km/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=kotlinpoet-km
-POM_NAME=KotlinPoet (Kotlin Metadata Extensions)
-POM_DESCRIPTION=Extensions for reading Kotlin metadata from Metadata annotations.
-POM_PACKAGING=jar
