+++ kotlinpoet-metadata-specs-tests/build.gradle.kts
@@ -13,41 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import org.jetbrains.dokka.gradle.DokkaTask
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
   kotlin("jvm")
-  id("org.jetbrains.dokka")
-  id("com.vanniktech.maven.publish") version versions.mavenPublish
 }
 
-val GROUP: String by project
-val VERSION_NAME: String by project
-
-group = GROUP
-version = VERSION_NAME
-
-tasks.named<Jar>("jar") {
-  manifest {
-    attributes("Automatic-Module-Name" to "com.squareup.kotlinpoet")
-  }
-}
-
-afterEvaluate {
-  tasks.named<DokkaTask>("dokka") {
-    skipDeprecated = true
-    outputDirectory = "$rootDir/docs/1.x"
-    outputFormat = "gfm"
-  }
+val compileTestKotlin: KotlinCompile by tasks
+compileTestKotlin.kotlinOptions {
+  freeCompilerArgs = listOf("-XXLanguage:+InlineClasses", "-Xjvm-default=enable")
+  jvmTarget = "1.8"
 }
 
 dependencies {
-  api(rootProject.project("kotlinpoet-metadata"))
-  api(deps.kotlin.stdlib)
-  implementation(deps.kotlin.reflect)
+  testImplementation(deps.kotlin.stdlib)
+  testImplementation(project(":kotlinpoet"))
+  testImplementation(project(":kotlinpoet-metadata"))
+  testImplementation(project(":kotlinpoet-metadata-specs"))
+  testImplementation(project(":kotlinpoet-elementhandler-elements"))
+  testImplementation(project(":kotlinpoet-elementhandler-reflective"))
   testImplementation(deps.kotlin.junit)
   testImplementation(deps.test.truth)
   testImplementation(deps.test.compileTesting)
-  testImplementation(deps.test.jimfs)
-  testImplementation(deps.test.ecj)
 }
