+++ kotlinpoet-metadata-specs/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=kotlinpoet-metadata-specs
+POM_NAME=KotlinPoet (KotlinPoet-metadata specs interop)
+POM_DESCRIPTION=Extensions for generating KotlinPoet specs from KotlinPoet-metadata types.
+POM_PACKAGING=jar
