+++ kotlinpoet-metadata/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=kotlinpoet-metadata
+POM_NAME=KotlinPoet (Kotlin Metadata Extensions)
+POM_DESCRIPTION=Extensions for reading Kotlin metadata from Metadata annotations.
+POM_PACKAGING=jar
